#74 vision api 관련 상태 저장 serice logic 개발
